+++ packages/client/src/components/Forum/index.ts
@@ -1,5 +1,4 @@
 export * from './TopicItem';
-export * from './NewTopicButton';
 export * from './Comment';
 export * from './CommentsList';
 export * from './NewCommentButton';
