+++ packages/client/src/constants/mockData.ts
@@ -15,51 +15,52 @@ export const mockForumPage = [
   },
 ];
 
-export const mockTopicPage = {
+export const mockTopic = {
   id: 111,
   title: 'Title of very interesting topic',
   content:
     'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
   author: 'Amanda',
-  comments: [
-    {
-      id: 1,
-      author: 'Jane',
-      content: 'First interesting comment',
-    },
-    {
-      id: 2,
-      author: 'David',
-      content:
-        'Second interesting comment (Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.)',
-    },
-    {
-      id: 3,
-      author: 'Jane',
-      content: 'First interesting comment',
-    },
-    {
-      id: 4,
-      author: 'David',
-      content:
-        'Second interesting comment (Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.)',
-    },
-    {
-      id: 5,
-      author: 'Jane',
-      content: 'First interesting comment',
-      parentCommentId: 2,
-    },
-    {
-      id: 6,
-      author: 'David',
-      content:
-        'Second interesting comment (Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.)',
-    },
-    {
-      id: 7,
-      author: 'Kate',
-      content: 'Third interesting comment',
-    },
-  ],
 };
+
+export const mockComments = [
+  {
+    id: 1,
+    author: 'Jane',
+    content: 'First interesting comment',
+  },
+  {
+    id: 2,
+    author: 'David',
+    content:
+      'Second interesting comment (Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.)',
+  },
+  {
+    id: 3,
+    author: 'Jane',
+    content: 'First interesting comment',
+  },
+  {
+    id: 4,
+    author: 'David',
+    content:
+      'Second interesting comment (Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.)',
+  },
+  {
+    id: 5,
+    author: 'Jane',
+    content: 'First interesting comment',
+    parentCommentId: 2,
+  },
+  {
+    id: 6,
+    author: 'David',
+    content:
+      'Second interesting comment (Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.)',
+  },
+  {
+    id: 7,
+    author: 'Kate',
+    content: 'Third interesting comment',
+  },
+];
